able to loop through multiple feeds
